+++ scripts/macros/2014/classFeatures/illrigger/interdictBoons/eyesOfTheGate.js
@@ -5,8 +5,8 @@ async function added({trigger: {entity: item}}) {
         1: 'interdictBoons'
     });
 }
-export let interdictBoonHellishFrenzy = {
-    name: 'Interdict Boons: Hellish Frenzy',
+export let interdictBoonEyesOfTheGate = {
+    name: 'Interdict Boons: Eyes of the Gate',
     version: '1.3.76',
     rules: 'legacy',
     item: [